frontend/requirements.txt
python-telegram-bot>=22.3
requests>=2.32.4
python-dotenv>=1.1.1